third_party/blink/web_tests/external/wpt/css/css-scroll-snap/snap-inline-block-expected.txt

This is a testharness.js-based test.
PASS Snaps correctly for horizontal-tb writing mode with 'scroll-snap-align: end start' alignment
PASS Snaps correctly for vertical-lr writing mode with 'scroll-snap-align: end start' alignment
PASS Snaps correctly for vertical-rl writing mode with 'scroll-snap-align: end start' alignment
PASS Snaps correctly for horizontal-tb writing mode with 'scroll-snap-align: start end' alignment
PASS Snaps correctly for vertical-lr writing mode with 'scroll-snap-align: start end' alignment
PASS Snaps correctly for vertical-rl writing mode with 'scroll-snap-align: start end' alignment
FAIL Snaps correctly for 'direction: rtl' with 'scroll-snap-align: end start' alignment assert_equals: aligns correctly on x expected -500 but got -315
FAIL Snaps correctly for 'direction: rtl' with 'scroll-snap-align: start end' alignment assert_equals: aligns correctly on x expected -315 but got -500
Harness: the test ran to completion.
